session-2/solution/requirements.txt
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
torch==1.10.0
ray==1.0.1.post1
pandas==1.3.4
torchvision==0.11.1